INFRA-102: Instance event job and Pipeline3 created in Jenkins.
All stages are in, but only the first two are implemented.
  new file:   jenkins/jenkins_home/jobs/pipeline3/config.xml
  new file:   jobs/pipelines/pipeline3.jenkinsfile
  new file:   app/spec/pipeline3/pipeline3.spec.js

New instance event job.
  new file:   jenkins/jenkins_home/jobs/instance-event/config.xml

Attempting to pass the instance name to the downstream job (even when it is unset).
Also checking the job config parameters.
  modified:   app/src/instance-event/validate-instance.js
  modified:   app/spec/instance-event/validate-instance.spec.js

Added the `set -e` flag to ensure that scripts fail on any of their failing command(s).
  modified:   app/src/pipeline3/host-preparation.js
  modified:   app/src/pipeline3/prehost-preparation.js

Added a bunch of new paths and variables.
  modified:   app/src/utils/config.js

Installing `rsync`.
  modified:   docker/Dockerfile

Improved the job and variables descriptions.
  modified:   jenkins/jenkins_home/jobs/pipeline1/config.xml

Adapted the 'host' section to follow our latest design.
  modified:   app/spec/instance-event/resources/test_instance_definition_2.json